feat: add responsive to navbar with loggedUser
