+++ src/components/index.ts
@@ -8,3 +8,4 @@ export * from './Logo/Logo';
 export * from './TwitchLive/TwitchLive';
 export * from './Tag/Tag';
 export * from './Countdown/Countdown';
+export * from './Avatar/Avatar';
